spatial-pathfinder/memecoin_bot/requirements.txt
requests
python-dotenv
solana
solders
websockets
